requirements.txt
# Cloth Recognition System - YOLOv8

# Core ML Dependencies
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Backend API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Gradio for quick demo
gradio>=4.0.0

# Utilities
roboflow>=1.1.0
tqdm>=4.66.0
pyyaml>=6.0.0
requests>=2.31.0
